Fill-a-pix-flask/requirements.txt
asgiref==3.3.4
click==7.1.2
colorama==0.4.4
Django==3.2
Flask==1.1.2
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
